requirements.txt
# Production Requirements
# Core Framework
flask==3.0.0
werkzeug==3.0.1

# Database & ORM
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
alembic==1.13.0

# API & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Security & CORS
flask-cors==4.0.0
python-dotenv==1.0.0

# RAG Pipeline
langchain==0.1.0
langchain-groq==0.0.1
langchain-core==0.1.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1

# ML/AI Dependencies
transformers==4.36.0
torch==2.1.2
numpy==1.26.2

# Utilities
python-multipart==0.0.6

# Production Server (WSGI)
gunicorn==21.2.0

# Monitoring & Logging
prometheus-client==0.19.0

pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
faker==20.1.0
httpx==0.25.2

# Code Quality
black==23.12.1
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
isort==5.13.2
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development Tools
ipython==8.18.1
ipdb==0.13.13
watchdog==3.0.0

# Type Stubs
types-requests==2.31.0.10
sqlalchemy-stubs==0.4
